[#159] fix: 반복 종료일 기본값을 anchor 로 + 경고 정리
- 사용자 리포트: '9월30일까지 반복' 설정했는데 실제로는
  9월30일부터 쭉 반복됨.
  원인: EndDateWheels 의 defaultEnd 가 anchor+1year 였음. 사용자가
  wheel 에서 월/일만 스크롤하고 연도 (2027 default) 는 놓쳐서
  365개짜리 시리즈가 생성됐던 것.
- defaultEnd = anchorDate (=시작일) 로 변경. 명시적으로 스크롤
  안 하면 1일짜리 시리즈만 생겨 사고 방지.

경고 정리:
- .dayOfMonth → .day (kotlinx-datetime 신 API)
- Share.android/ios @ParameterName annotation 일치
- MainScreen.plusMonths 의 중복 .toInt() 제거
- shared build 에 -Xexpect-actual-classes 플래그로 Beta 경고 억제
- AnniversariesScreen 의 손상된 확장 함수 복구 (직전 batch 처리
  실수로 .month.number 로 잘못 변환됐던 것)

남은 경고:
- monthNumber deprecation — kotlinx-datetime 0.8 에는 대체할
  Month.number 가 아직 없음 (0.9+ 예정). 그때 처리.
- LocalClipboardManager deprecation — LocalClipboard 는 suspend
  API 라 별도 리팩터 필요, 별도 이슈로.

+++ shared/src/commonMain/kotlin/com/hyunjine/linker/feature/main/MainScreen.kt
@@ -139,7 +139,7 @@ data class YearMonth(val year: Int, val month: Int) {
     fun plusMonths(months: Int): YearMonth {
         val total = year * 12L + (month - 1) + months
         val newYear = total.floorDiv(12).toInt()
-        val newMonth = total.mod(12).toInt() + 1
+        val newMonth = total.mod(12) + 1
         return YearMonth(newYear, newMonth)
     }
